CI: Update binaries cache
Only cache binaries if source files have not changed. This is helpful if one is
iterating on CI and wants to avoid rebuilding the Haskell source and binaries.

Noticed the existing cache was too aggressive when I reformatted the code and
the build was skipped because `.cabal`, `stack.yaml`, etc. files didn’t
change.